+++ include/rain/networking/http/client.hpp
@@ -1,0 +1,8 @@
+// TODO: HTTP client.
+#pragma once
+
+#include "socket.hpp"
+
+namespace Rain::Networking::Http {
+	class Client : protected Http::Socket {};
+}